chore: move granules index to correct location
