+++ cumulus_lambda_functions/lib/uds_db/granules_db_index.py
